ボタン設定を別ファイルに変更
